requirements.txt
alembic==1.14.0
blinker==1.9.0
Bootstrap-Flask==2.3.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dominate==2.9.1
Flask==2.3.3
Flask-Limiter==3.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.1.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.18.0
python-dotenv==1.0.0
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.0.1

# Google Cloud Vision API
google-cloud-vision==3.4.4
Pillow==10.0.0

# OpenAI API
openai==1.12.0

# Development tools
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Production tools
sentry-sdk==1.40.0
redis==5.0.1
celery==5.3.6

# Performance optimization
Flask-Caching==2.3.0
Flask-Compress==1.14
flask-limiter==3.5.0

# Security
Flask-Talisman==1.1.0
cryptography==41.0.7

# Monitoring
prometheus-flask-exporter==0.23.0
